+++ lib/ui/widgets/donation_countdown_timer.dart
@@ -1,11 +1,10 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:zachranobed/common/constants.dart';
 import 'package:zachranobed/common/helper_service.dart';
+import 'package:zachranobed/common/lifecycle/lifecycle_watcher.dart';
 import 'package:zachranobed/models/canteen.dart';
 import 'package:zachranobed/models/delivery.dart';
 import 'package:zachranobed/notifiers/delivery_notifier.dart';
@@ -17,31 +16,76 @@ class DonationCountdownTimer extends StatefulWidget {
   State<DonationCountdownTimer> createState() => _DonationCountdownTimerState();
 }
 
-class _DonationCountdownTimerState extends State<DonationCountdownTimer> {
-  Timer? _countdownTimer;
-  Duration _remainingTime = const Duration(seconds: 0);
+class _DonationCountdownTimerState extends State<DonationCountdownTimer>
+    with TickerProviderStateMixin, LifecycleWatcher {
+  Ticker? _ticker;
+  String _value = "";
 
   @override
   void initState() {
     super.initState();
-    _remainingTime = _getRemainingTimeForDonation();
-    if (_remainingTime.inSeconds > 0) {
-      _startTimer();
-    }
+
+    _checkTicker();
   }
 
   @override
   void dispose() {
-    if (_countdownTimer != null) {
-      _countdownTimer!.cancel();
-    }
+    _stopTicker();
     super.dispose();
+  }
+
+  @override
+  void onResume() {
+    _checkTicker();
+  }
+
+  void _checkTicker() {
+    final total = _getRemainingTimeForDonation();
+    if (total.inSeconds > 0) {
+      _startTicker(total);
+    } else {
+      _stopTicker();
+    }
+  }
+
+  void _startTicker(Duration total) {
+    _ticker?.dispose();
+    _ticker = createTicker((elapsed) => _countdownTick(total, elapsed));
+    _ticker?.start();
+  }
+
+  void _stopTicker() {
+    if (_ticker != null) {
+      _ticker?.dispose();
+
+      SchedulerBinding.instance.addPostFrameCallback((_) {
+        context
+            .read<DeliveryNotifier>()
+            .updateDeliveryState(DeliveryState.notUsed);
+      });
+    }
+  }
+
+  void _countdownTick(Duration total, Duration elapsed) {
+    final time = total - elapsed;
+    final hours = time.inHours.remainder(24).toString().padLeft(2, '0');
+    final minutes = time.inMinutes.remainder(60).toString().padLeft(2, '0');
+    final seconds = time.inSeconds.remainder(60).toString().padLeft(2, '0');
+
+    final newValue = '$hours:$minutes:$seconds';
+    if (newValue != _value) {
+      setState(() => _value = newValue);
+    }
+
+    if (elapsed > total) {
+      _stopTicker();
+    }
   }
 
   Duration _getRemainingTimeForDonation() {
     final user = HelperService.getCurrentUser(context) as Canteen;
     String timeNow = DateFormat('HH:mm:ss').format(DateTime.now());
-    String donateWithin = user.pickUpFrom!;
+    String donateWithin = user.pickUpFrom;
 
     DateTime startTime = DateFormat('HH:mm:ss').parse(timeNow);
     DateTime endTime = DateFormat('HH:mm').parse(donateWithin);
@@ -50,46 +94,10 @@ class _DonationCountdownTimerState extends State<DonationCountdownTimer> {
         const Duration(minutes: Constants.pickupConfirmationTime);
   }
 
-  void _startTimer() {
-    _countdownTimer =
-        Timer.periodic(const Duration(seconds: 1), (_) => _setCountDown());
-  }
-
-  void _setCountDown() {
-    const reduceSecondsBy = 1;
-    if (mounted) {
-      setState(() {
-        final seconds = _remainingTime.inSeconds - reduceSecondsBy;
-        if (seconds < 0) {
-          _countdownTimer!.cancel();
-        } else {
-          _remainingTime = Duration(seconds: seconds);
-        }
-      });
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
-    final hours =
-        _remainingTime.inHours.remainder(24).toString().padLeft(2, '0');
-    final minutes =
-        _remainingTime.inMinutes.remainder(60).toString().padLeft(2, '0');
-    final seconds =
-        _remainingTime.inSeconds.remainder(60).toString().padLeft(2, '0');
-
-    final canDonate = _countdownTimer?.isActive ?? false;
-
-    if (!canDonate) {
-      SchedulerBinding.instance.addPostFrameCallback((_) {
-        context.read<DeliveryNotifier>().updateDeliveryState(
-          DeliveryState.notUsed
-        );
-      });
-    }
-
     return Text(
-      '$hours:$minutes:$seconds',
+      _value,
       style: const TextStyle(
         color: ZOColors.onPrimaryLight,
         fontSize: FontSize.s,